Move numpy tests back to main file.
